renamed to Alooma-iOS, updated README.md
also cleaned up project file
