+++ Alooma-iOS/AloomaLogger.h
