+++ Alooma-iOS/NSData+AloomaBase64.h
