+++ Alooma-iOS/NSData+AloomaBase64.m
